// src/http-pref-rio-carta-servicos/models/channelType.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API Carta de Serviços - Pref.Rio
 * API pública para consumo da Carta de Serviços Municipais do Rio de Janeiro, integrada com Salesforce.
 * OpenAPI spec version: 1.0.1
 */

export type ChannelType = (typeof ChannelType)[keyof typeof ChannelType]

export const ChannelType = {
  Presencial: 'Presencial',
  Digital: 'Digital',
  Telefone: 'Telefone',
} as const